requirements.txt
watchdog>=4.0.0
pystray>=0.19.0
Pillow>=10.0.0
psutil>=5.9.0
pyinstaller>=6.0.0
